docs(readme): fix markdown formatting and update CLI commands in README.md
- Correct markdown syntax for horizontal rules and list items for better readability
- Update example CLI commands to new module paths and names for Architect, Chief Engineer, and Director roles
- Ensure consistent code block formatting for commands and sample configuration
- Preserve original content context and improve overall structure without content changes

+++ src/backend/polaris/kernelone/benchmark/holographic_stats.py
@@ -52,7 +52,7 @@ def percentile(samples: list[float], ratio: float) -> float:
         return 0.0
     ratio = min(max(ratio, 0.0), 1.0)
     sorted_values = sorted(samples)
-    index = int(math.ceil(ratio * len(sorted_values))) - 1
+    index = math.ceil(ratio * len(sorted_values)) - 1
     index = min(max(index, 0), len(sorted_values) - 1)
     return float(sorted_values[index])
